api/requirements.txt
Flask==1.0.2
werkzeug==0.16.0
onnxruntime==1.0.0
Pillow==9.5.0
numpy==1.21.6